+++ src/App.vue
@@ -1,17 +1,17 @@
 <template>
   <div id="app">
     <img src="./assets/logo.png">
-    <HelloWorld msg="Welcome to Your Vue.js App"/>
+    <Flappy msg="Welcome to Your Vue.js App"/>
   </div>
 </template>
 
 <script>
-import HelloWorld from './components/HelloWorld.vue'
+import Flappy from './components/Flappy.vue'
 
 export default {
   name: 'app',
   components: {
-    HelloWorld
+    Flappy
   }
 }
 </script>
